bhashamitra-backend/apps/core/pagination.py
"""Custom pagination classes."""
from rest_framework.pagination import CursorPagination as BaseCursorPagination


class CursorPagination(BaseCursorPagination):
    """Default cursor pagination."""
    page_size = 20
    page_size_query_param = 'limit'
    max_page_size = 50
    ordering = '-created_at'
